move inverter device testing to separate test file and reduce code lines
